Changed cost function of neural network to cross entropy, developed math algorithm, converted neural network to a class, and utilized it in a model.
The cost function was originally mean squared error, but cross entropy provided better results. Math algorithm to determine the value from a list of strings of numbers and symbols has been finished. Extensive testing has not been done yet. Parenthesis to be implemented later, but the four basic operators are functional.

Neural network has been converted to a class and the object is used in model.py to train and test.